third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-added-display-none-host-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests adding a shadow root to a display none host doesn't leave dirty bits.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inside).borderWidth is "0px"
PASS getComputedStyle(inside).borderWidth is "10px"
PASS successfullyParsed is true

TEST COMPLETE
